pyproject.toml
[tool.poetry]
name = "coinmetrics-api-client"
version = "2023.8.25.15"
description = "Python client for Coin Metrics API v4."
authors = ["Coin Metrics <info@coinmetrics.io>", "Oleksandr Buchkovskyi <oleksandr@coinmetrics.io>"]
license = "MIT"
readme = "README.md"
homepage = "https://coinmetrics.github.io/api-client-python/site/index.html"
repository = "https://github.com/coinmetrics/api-client-python"
packages = [{include = "coinmetrics"}]
keywords = ["coin metrics", "coin", "metrics", "crypto", "bitcoin", "network-data", "market-data", "api", "handy"]
documentation = "https://coinmetrics.github.io/api-client-python/site/index.html"
include = ["test"]

[tool.poetry.dependencies]
python = "^3.7.1"
requests = "^2.24.0"
orjson = "^3.6.0"
pandas = {version="^1.3.3", optional=true}
websocket-client = "^1.2.1"
python-dateutil = "^2.8.2"
typer = "^0.7.0"
tqdm = "^4.64.1"

[tool.poetry.dev-dependencies]
mypy = "^1.1.1"
flake8 = "^3.8.3"
black = "^23.1"
mkdocs = "^1.3.1"
mkdocs-material = "^9.1.3"
pydoc-markdown = "^4.8"

pytest-mock = "^3.2.0"

[tool.poetry.scripts]
coinmetrics = 'coinmetrics.typer_cli:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
pandas = ["pandas"]
